Add a common base test class
To make setups easier across all test classes, a base class is written for the common functionality.

This base class automatically creates mocks for all dependencies, but still allows client to configure if needed.

Also, it removes the need of manually creating test instances in each method.

PS: This is a simplified version of the implementation in Armut, therefore all credits go to the creative backend engineers in there.

PS2: I don't know whether it is something trivial or not due to me being absolute noob in backend unit testing but still I had quite a lot of fun reading the original code. A lovely implementation really.