Add auto-migrate functionality
In order for this API to function normally, it needs to connect to a DB that is setup correctly with necessary tables.

This is normally done by running EF Core migrations against the DB.

Whilst developing this app, these migrations are ran manually on localhost by using CLI. However, when this API and DB are dockerized, it became impossible to migrate manually because API and DB run on different containers.

To fix this issue, a small script is added to Program.cs that controls whether the DB has necessary EF migrations or not.

If there are no migrations applied, it applies the migrations.
If migrations are already applied, it skips doing migrations twice.